+++ app/src/main/java/woowacourse/payments/ui/screen/cardAddition/CardAdditionActivity.kt
@@ -1,12 +1,12 @@
-package woowacourse.payments
+package woowacourse.payments.ui.screen.cardAddition
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import woowacourse.payments.ui.CardAdditionScreen
+import woowacourse.payments.ui.screen.cardAddition.component.CardAdditionScreen
 
-class MainActivity : ComponentActivity() {
+class CardAdditionActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
