+++ app/src/main/java/woowacourse/payments/ui/screen/cardAddition/component/PaymentCard.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.screen.cardAddition.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
